try to fix pixel blurring, and clamp positions to ints

+++ examples/game_2d/src/lighting.hpp
@@ -14,7 +14,7 @@ namespace game2d {
 
 void
 generate_intersections(GameObject2D& camera,
-                       glm::vec2& light_pos,
+                       glm::ivec2& light_pos,
                        const std::vector<std::reference_wrapper<GameObject2D>>& ents,
                        const glm::ivec2& screen_wh,
                        std::vector<std::tuple<float, float, float>>& intersections);
